Star Rating Captured
The star ratings are now captured for the ‘Film’ entity, and the
relevant numbered star images are displayed on the table cells.

The selected film image is now displayed on the background of the
DetailVC (once it has been selected).

Fixed the form validation to include checking if the star ratings have
been set as well as the fields populated. Setting an image has been
intentionally left as optional (in case the user doesn’t have an image)
but this may be revisited.