Edit Functionality Added
It is now possible for the user to edit all of a film’s details by
viewing the details (tapping its table view cell), and then selecting
the ‘Edit’ button.

Also added the film image to be part of the form validation (it checks
the image has been changed and has not been set as the placeholder
image).

Added ‘.xcbkptlist’ files to the ‘.gitignore’ file.